+++ RedPandaIDE/compiler/compiler.cpp
@@ -423,11 +423,6 @@ QString Compiler::getCCompileArguments(bool checkSyntax)
             foreach(const QString& param, params)
                 result += " "+ parseMacros(param);
         }
-    } else {
-        if (compilerSet()->maxFrameSize()>0 && compilerSet()->warnLargeFrame()) {
-            long long size = std::ceil(compilerSet()->maxFrameSize()*1024*1024);
-            result += QString(" -Werror=frame-larger-than=%1").arg(size);
-        }
     }
     return result;
 }
@@ -467,11 +462,6 @@ QString Compiler::getCppCompileArguments(bool checkSyntax)
             QStringList params = textToLines(s);
             foreach(const QString& param, params)
                 result += " "+ parseMacros(param);
-        }
-    } else {
-        if (compilerSet()->maxFrameSize()>0 && compilerSet()->warnLargeFrame()) {
-            long long size = std::ceil(compilerSet()->maxFrameSize()*1024*1024);
-            result += QString(" -Werror=frame-larger-than=%1").arg(size);
         }
     }
     return result;