+++ RedPandaIDE/settingsdialog/compilersetoptionwidget.cpp
@@ -88,8 +88,6 @@ static void loadCompilerSetSettings(Settings::PCompilerSet pSet, Ui::CompilerSet
     ui->txtCustomLinkParams->setEnabled(pSet->useCustomLinkParams());
     ui->chkAutoAddCharset->setChecked(pSet->autoAddCharsetParams());
     ui->chkStaticLink->setChecked(pSet->staticLink());
-    ui->spinMaxObjectSize->setValue(pSet->maxFrameSize());
-    ui->chkWarnLargeObject->setChecked(pSet->warnLargeFrame());
     //rest tabs in the options widget
 
     ui->optionTabs->resetUI(pSet,pSet->compileOptions());
@@ -201,8 +199,6 @@ void CompilerSetOptionWidget::saveCurrentCompilerSet()
     pSet->setCustomLinkParams(ui->txtCustomLinkParams->toPlainText().trimmed());
     pSet->setAutoAddCharsetParams(ui->chkAutoAddCharset->isChecked());
     pSet->setStaticLink(ui->chkStaticLink->isChecked());
-    pSet->setMaxFrameSize(ui->spinMaxObjectSize->value());
-    pSet->setWarnLargeFrame(ui->chkWarnLargeObject->isChecked());
 
     pSet->setCCompiler(ui->txtCCompiler->text().trimmed());
     pSet->setCppCompiler(ui->txtCppCompiler->text().trimmed());